app/requirements.txt
Django==1.8.6
django-nose==1.4.2
djangorestframework==3.3.1
Flask==0.10.1
httpie==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
nose==1.3.7
passlib==1.6.5
Pygments==2.0.2
requests==2.8.1
vboxapi==1.0
Werkzeug==0.11.2
wheel==0.26.0
